Create Vengance_Paladin_Rick
This is a mock-up of how a character can be built and generate damage for use in "gaming Rick". This will be a larger project for the citadel of Ricks that I'll pursue in the near future. This program actually generates mean damage per round for my vengeance oath paladin level 5 from D&D 5e based on the combat rounds desired, AC of the enemy, and my strength bonus.